requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.3.0
perplexity==0.0.8
gunicorn==21.2.0